remover modules of pages ned to remove layout and core modules in the future
