+++ package.json
@@ -53,6 +53,14 @@
         "yarn": "^1.22.19",
         "zone.js": "^0.13.3"
     },
+    "overrides": {
+        "@angular/fire": {
+            "@angular/common": "^16.0.0",
+            "@angular/core": "^16.0.0",
+            "@angular/platform-browser": "^16.0.0",
+            "@angular/platform-browser-dynamic": "^16.0.0"
+        }
+    },
     "devDependencies": {
         "@angular-devkit/build-angular": "^16.2.5",
         "@angular-eslint/builder": "16.2.0",
